v3 refactored AbstractResource up one level and made rights an array-only key
