+++ lib/iiif/v3/presentation/annotation_collection.rb
@@ -1,9 +1,7 @@
-require File.join(File.dirname(__FILE__), 'abstract_resource')
-
 module IIIF
   module V3
     module Presentation
-      class AnnotationCollection < AbstractResource
+      class AnnotationCollection < IIIF::V3::AbstractResource
 
         TYPE = 'AnnotationCollection'
