+++ lib/iiif/v3/presentation/annotation_page.rb
@@ -1,9 +1,7 @@
-require File.join(File.dirname(__FILE__), 'abstract_resource')
-
 module IIIF
   module V3
     module Presentation
-      class AnnotationPage < AbstractResource
+      class AnnotationPage < IIIF::V3::AbstractResource
 
         TYPE = 'AnnotationPage'
