+++ lib/iiif/v3/presentation/image_resource.rb
@@ -1,4 +1,3 @@
-require File.join(File.dirname(__FILE__), 'resource')
 require 'faraday'
 require 'json'
