+++ lib/iiif/v3/presentation/manifest.rb
@@ -1,9 +1,7 @@
-require File.join(File.dirname(__FILE__), 'abstract_resource')
-
 module IIIF
   module V3
     module Presentation
-      class Manifest < AbstractResource
+      class Manifest < IIIF::V3::AbstractResource
 
         TYPE = 'Manifest'
