+++ lib/iiif/v3/presentation/resource.rb
@@ -1,9 +1,7 @@
-require File.join(File.dirname(__FILE__), 'abstract_resource')
-
 module IIIF
   module V3
     module Presentation
-      class Resource < AbstractResource
+      class Resource < IIIF::V3::AbstractResource
 
         def required_keys
           %w{ id }
