+++ lib/iiif/v3/presentation/sequence.rb
@@ -1,9 +1,7 @@
-require File.join(File.dirname(__FILE__), 'abstract_resource')
-
 module IIIF
   module V3
     module Presentation
-      class Sequence < AbstractResource
+      class Sequence < IIIF::V3::AbstractResource
 
         TYPE = 'Sequence'
